SharedFoldersSvc: Added new function for server-side file content copying (SHFL_FN_COPY_FILE_PART) and a function for query host features and supported functions (SHFL_FN_QUERY_FEATURES). [build fix]  bugref:9172
svn:sync-xref-src-repo-rev: r129521

+++ src/VBox/HostServices/SharedFolders/vbsf.cpp
@@ -1,4 +1,4 @@
-/* $Id: vbsf.cpp 77837 2019-03-21 21:32:32Z knut.osmundsen@oracle.com $ */
+/* $Id: vbsf.cpp 77838 2019-03-21 21:50:38Z knut.osmundsen@oracle.com $ */
 /** @file
  * Shared Folders - VBox Shared Folders.
  */
@@ -1381,7 +1381,7 @@ int vbsfCopyFilePart(SHFLCLIENTDATA *pClient, SHFLROOT idRootSrc, SHFLHANDLE hFi
 
     AssertPtrReturn(pClient, VERR_INVALID_PARAMETER);
 
-    size_t cbTotal = 0;
+    uint64_t cbTotal = 0;
 
     SHFLFILEHANDLE *pHandleSrc = vbsfQueryFileHandle(pClient, hFileSrc);
     int rc = vbsfCheckHandleAccess(pClient, idRootSrc, pHandleSrc, VBSF_CHECK_ACCESS_READ);
